% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{hist_box_plots}
\alias{hist_box_plots}
\title{Side by Side Histogram and Box Plots}
\usage{
hist_box_plots(dat, x)
}
\arguments{
\item{dat}{The dataset}

\item{x}{The variable in the dataset}
}
\value{
nothing
}
\description{
Side by Side Histogram and Box Plots
}
\examples{
nothing
}
