% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{stacked_bar_plots}
\alias{stacked_bar_plots}
\title{Stacked Histograms}
\usage{
stacked_bar_plots(dat_a, dat_b, x)
}
\arguments{
\item{dat_a}{The dataset for stack 1}

\item{dat_b}{The dataset for stack 2}

\item{x}{The variable in the dataset}
}
\value{
nothing
}
\description{
Stacked Histograms
}
\examples{
nothing
}
